% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stern_brocot_tree.R
\name{plot_stern_brocot_tree}
\alias{plot_stern_brocot_tree}
\title{Plot Stern-Brocot Tree}
\usage{
plot_stern_brocot_tree(path)
}
\arguments{
\item{path}{A binary string (e.g., "1010") representing the traversed nodes.}
}
\value{
A plot of the Stern-Brocot tree with the traversed nodes highlighted.
}
\description{
Plot Stern-Brocot Tree
}
\examples{
plot_stern_brocot_tree("1010")
}
