% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stern_brocot}
\alias{stern_brocot}
\title{stern_brocot}
\usage{
stern_brocot(x, lower_uncertainty, upper_uncertainty)
}
\arguments{
\item{x}{Vector of numbers to convert to rational fractions.}

\item{lower_uncertainty}{Lower uncertainty bounds (scalar or vector).}

\item{upper_uncertainty}{Upper uncertainty bounds (scalar or vector).}
}
\value{
A data frame with results and metadata about the tree path.
}
\description{
Approximates a floating-point number to arbitrary uncertainty (vectorized version).
}
\details{
This implementation is based on the Stern-Brocot tree algorithm as described in:
\itemize{
\item Forišek M. Approximating rational numbers by fractions. In International Conference on Fun with Algorithms 2007 Jun 3 (pp. 156-165). Berlin, Heidelberg: Springer Berlin Heidelberg.
\item Stolzenburg F. Harmony perception by periodicity detection. Journal of Mathematics and Music. 2015 Sep 2;9(3):215-38.
}
}
